@@ -12,36 +12,38 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.0.0",
-    "@angular/common": "^4.0.0",
-    "@angular/compiler": "^4.0.0",
-    "@angular/core": "^4.0.0",
-    "@angular/forms": "^4.0.0",
-    "@angular/http": "^4.0.0",
-    "@angular/platform-browser": "^4.0.0",
-    "@angular/platform-browser-dynamic": "^4.0.0",
-    "@angular/router": "^4.0.0",
+    "@angular/animations": "7.0.0",
+    "@angular/common": "7.0.0",
+    "@angular/compiler": "7.0.0",
+    "@angular/core": "7.0.0",
+    "@angular/forms": "7.0.0",
+    "@angular/http": "7.0.0",
+    "@angular/platform-browser": "7.0.0",
+    "@angular/platform-browser-dynamic": "7.0.0",
+    "@angular/router": "7.0.0",
     "@fortawesome/fontawesome-free": "^5.3.1",
     "@fortawesome/fontawesome-free-webfonts": "^1.0.4",
     "@types/lodash": "^4.14.106",
-    "angularfire2": "^5.0.0-rc.3",
     "bulma": "^0.7.1",
     "bulma-divider": "^2.0.1",
     "core-js": "^2.5.5",
-    "firebase": "^4.7.0",
+    "@angular/fire": "^5.0.2",
+    "firebase": "^5.0.3",
     "firebase-admin": "^6.0.0",
     "firebase-functions": "^2.0.5",
     "hammerjs": "^2.0.8",
     "lodash-es": "^4.17.8",
-    "ngx-toastr": "6.4.1-beta.0",
+    "ngx-toastr": "9.1.1",
     "nodemailer": "^4.6.4",
-    "rxjs": "^5.5.9",
+    "rxjs": "^6.3.3",
+    "rxjs-tslint": "^0.1.5",
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "^1.2.1",
-    "@angular/compiler-cli": "^4.0.0",
-    "@angular/language-service": "^4.0.0",
+    "@angular-devkit/build-angular": "~0.10.0",
+    "@angular/cli": "^7.0.2",
+    "@angular/compiler-cli": "7.0.0",
+    "@angular/language-service": "7.0.0",
     "@types/jasmine": "~2.5.53",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "^6.0.105",
@@ -58,6 +60,6 @@
     "protractor": "~5.1.2",
     "ts-node": "~3.0.4",
     "tslint": "~5.3.2",
-    "typescript": "~2.3.3"
+    "typescript": "3.1.3"
   }
 }
