@@ -3,12 +3,13 @@ import {AppComponent} from './app.component';
 import {BrowserModule} from "@angular/platform-browser";
 import {ToastrModule} from "ngx-toastr";
 import {BrowserAnimationsModule} from "@angular/platform-browser/animations";
-import {AngularFireDatabaseModule} from "angularfire2/database";
-import {AngularFireModule} from "angularfire2";
 import {environment} from "../environments/environment";
 import {SharedModule} from "./shared/shared.module";
 import {HomeModule} from "./home/home.module";
 import { GalleryComponent } from './gallery/gallery.component';
+import {AngularFireDatabaseModule} from "@angular/fire/database";
+import {AngularFireModule} from "@angular/fire";
+import {AngularFireFunctionsModule} from "@angular/fire/functions";
 
 @NgModule({
   declarations: [
@@ -21,6 +22,7 @@ import { GalleryComponent } from './gallery/gallery.component';
     ToastrModule.forRoot(),
     AngularFireDatabaseModule,
     AngularFireModule.initializeApp(environment.firebase),
+    AngularFireFunctionsModule,
     HomeModule,
     SharedModule.forRoot()
   ],
